% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getabstract.R
\name{getabstract}
\alias{getabstract}
\title{This script is for a function to take a doi input and scrape back an abstract from the
publisher's website based on the presence of a subheading and the layout of the text.}
\usage{
getabstract(doi = NULL)
}
\arguments{
\item{doi}{A string of length 1 containing the doi for a journal article}
}
\value{
A string containing the abstract text. May include some other publisher text, including citation information and some background text where abstracts cannot be parsed easily.
}
\description{
Scrape abstract from publisher's website.
}
